Initial Push - Very early stages
At this point, Dimensions are tracked at compile time.

Units are tracked at runtime.

Dimensions can be simplified, but units will not
yet be handled appropriately.

Division and Multiplication of Dimensions works.

Unit test structure is in place,
but unit tests are not yet in use.

Apache 2.0 license is added.